split virtual machine tests
